pyproject.toml
[tool.poetry]
name = "claude-chatbot"
version = "0.1.0"
description = "A terminal-based chatbot using Claude 3.7 LLM"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[[tool.poetry.source]]
name = "sarvam-python-ci"
url = "https://europe-west4-python.pkg.dev/gpu-reservation-sarvam/sarvam-python-ci/simple/"
priority = "explicit"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.4"
litellm = "^1.30.0"
google-cloud-aiplatform = "^1.50.0"
python-dotenv = "^1.0.0"
rich = "^13.7.0"
click = "^8.1.7"
httpx = "0.27.2"
jinja2 = "^3.1.0"
sarvam-datatypes = {version = "0.4.20", source = "sarvam-python-ci"}
sarvam-stream = {version = "^0.3.9", source = "sarvam-python-ci"}
sarvam-agents-sdk = {version = "^0.2.171", source = "sarvam-python-ci"}
sqlalchemy = "^2.0.41"
streamlit = "^1.28.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.12.0"
flake8 = {version = "^6.1.0", python = ">=3.8.1,<4.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"] 